@@ -9,63 +9,65 @@
   },
   "dependencies": {
     "@riophae/vue-treeselect": "^0.4.0",
-    "@sentry/vue": "^6.2.1",
-    "@vue/composition-api": "^1.1.4",
-    "axios": "^0.21.2",
+    "@sentry/vue": "^6.16.1",
+    "@vue/composition-api": "^1.4.1",
+    "axios": "^0.24.0",
     "axios-cache-adapter": "^2.7.3",
-    "citation-js": "^0.5.0-alpha.4",
-    "core-js": "^3.4.4",
-    "d3": "^6.2.0",
-    "d3-array": "^2.8.0",
-    "d3-axis": "^2.0.0",
-    "d3-brush": "^2.1.0",
-    "d3-scale": "^3.2.3",
-    "d3-selection": "^2.0.0",
-    "d3-time": "^2.0.0",
-    "d3-time-format": "^3.0.0",
-    "filesize": "^6.0.1",
+    "citation-js": "^0.5.4",
+    "core-js": "^3.19.3",
+    "d3": "^7.2.1",
+    "d3-array": "^3.1.1",
+    "d3-axis": "^3.0.0",
+    "d3-brush": "^3.0.0",
+    "d3-scale": "^4.0.2",
+    "d3-selection": "^3.0.0",
+    "d3-time": "^3.0.0",
+    "d3-time-format": "^4.1.0",
+    "filesize": "^8.0.6",
     "leaflet": "^1.7.1",
-    "leaflet.markercluster": "^1.5.0",
+    "leaflet.markercluster": "^1.5.3",
     "lodash": "^4.17.21",
-    "moment": "^2.27.0",
+    "moment": "^2.29.1",
     "protobufjs": "^6.11.2",
-    "vue": "^2.6.10",
-    "vue-async-computed": "^3.8.2",
-    "vue-google-charts": "^0.3.2",
-    "vue-gtag": "^1.11.0",
-    "vue-router": "^3.3.4",
+    "serialize-javascript": "^6.0.0",
+    "vue": "^2.6.14",
+    "vue-async-computed": "^3.9.0",
+    "vue-google-charts": "^0.3.3",
+    "vue-gtag": "^1.16.1",
+    "vue-router": "^3.5.3",
     "vue2-leaflet": "^2.7.1",
     "vue2-leaflet-markercluster": "^3.1.0",
-    "vuetify": "^2.5.6"
+    "vuetify": "^2.6.1"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "@types/d3": "^5.16.3",
-    "@types/d3-array": "^2.0.0",
-    "@types/d3-brush": "^2.1.0",
-    "@types/d3-scale": "^2.2.0",
-    "@types/d3-selection": "^2.0.0",
-    "@types/d3-time-format": "^2.1.1",
-    "@types/leaflet": "^1.7.4",
-    "@types/lodash": "^4.14.159",
-    "@typescript-eslint/eslint-plugin": "^3.6.0",
-    "@typescript-eslint/parser": "^3.6.0",
-    "@vue/cli-plugin-babel": "^4.1.0",
-    "@vue/cli-plugin-eslint": "^4.4.0",
-    "@vue/cli-plugin-typescript": "^4.4.6",
-    "@vue/cli-service": "^4.1.0",
-    "@vue/eslint-config-airbnb": "^5.1.0",
-    "@vue/eslint-config-typescript": "^5.1.0",
+    "@types/d3": "^7.1.0",
+    "@types/d3-array": "^3.0.2",
+    "@types/d3-brush": "^3.0.1",
+    "@types/d3-scale": "^4.0.2",
+    "@types/d3-selection": "^3.0.2",
+    "@types/d3-time-format": "^4.0.0",
+    "@types/leaflet": "^1.7.6",
+    "@types/lodash": "^4.14.178",
+    "@typescript-eslint/eslint-plugin": "^5.6.0",
+    "@typescript-eslint/parser": "^5.6.0",
+    "@vue/cli-plugin-babel": "^4.5.15",
+    "@vue/cli-plugin-eslint": "^4.5.15",
+    "@vue/cli-plugin-typescript": "^4.5.15",
+    "@vue/cli-service": "^4.5.15",
+    "@vue/eslint-config-airbnb": "^6.0.0",
+    "@vue/eslint-config-typescript": "^9.1.0",
     "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "sass": "^1.26.3",
-    "sass-loader": "^8.0.2",
-    "typescript": "^4.0.3",
-    "vue-cli-plugin-vuetify": "^2.4.1",
-    "vue-template-compiler": "^2.6.10",
-    "vuetify-loader": "^1.7.2"
+    "eslint": "^7.32.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-vue": "^8.2.0",
+    "eslint-plugin-vuejs-accessibility": "^1.1.0",
+    "sass": "~1.32.0",
+    "sass-loader": "^10.1.1",
+    "typescript": "^4.5.3",
+    "vue-cli-plugin-vuetify": "^2.4.5",
+    "vue-template-compiler": "^2.6.14",
+    "vuetify-loader": "^1.7.3"
   },
   "browserslist": [
     "> 1%",
